% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{dualpathSvd2}
\alias{dualpathSvd2}
\title{From https://github.com/statsmaths/genlasso/blob/master/R/dualpathSvd.R
modified to compute the polyhedron post-selection inference selection.}
\usage{
dualpathSvd2(y, D, approx = FALSE, maxsteps = 2000, minlam = 0,
  rtol = 1e-07, btol = 1e-07, verbose = FALSE, object = NULL,
  ctol = 1e-10, cdtol = 1e-04, do.gc = F)
}
\arguments{
\item{y}{numeric vector of data.}

\item{D}{penalty matrix.}

\item{approx}{If approx=TRUE, then the fused lasso path will be computed
without any coordinates leaving the path.}

\item{maxsteps}{maximum number of steps of the algorithms to run.}

\item{rtol}{Tolerance for solving linear system using svdsolve(), defaults to 1e-7.}

\item{btol}{Tolerance for leaving times, precision issue, defaults to 1e-10}

\item{ctol}{Tolerance for gmat&%y ~= c; defaults to 1e-7}

\item{cdtol}{Tolerance for cdtol; defaults to 1e-4}
}
\description{
We compute a solution path of the generalized lasso dual problem:
% \hat{u}(\lambda) =
\argmin_u \|y - D^T u\|_2^2 \rm{s.t.} \|\u\|_\infty \leq \lambda %
}
\details{
where %D% is %m \times n%. Here there is no assumption on %D%, and we use a
fresh SVD at each iteration (computationally naive but stable).

Note: the df estimates at each %lambda_k% can be thought of as the df for all
solutions corresponding to lambda in %(lambda_k,lambda_{k-1})%, the open
interval to the *right* of the current %lambda_k%.
}

