% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{evec}
\alias{evec}
\title{creates vector of length n whose j'th element is 1, otherwise 0. Not currently in use.}
\usage{
evec(n, j)
}
\description{
creates vector of length n whose j'th element is 1, otherwise 0. Not currently in use.
}

