% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{get.modelinfo}
\alias{get.modelinfo}
\title{Calculates various model-related quantities for _any_ D for the signal
approximator case (IC, penalty, RSS, residual) \code{df.fun()} is a function that
returns the degrees of freedom of fit of %yhat = Proj_{null(D_B)} y%}
\usage{
get.modelinfo(f0, y0, sigma, maxsteps, D, stoprule = c("bic", "ebic", "aic"),
  ebic.fac = 0.5, verbose = F)
}
\description{
Calculates various model-related quantities for _any_ D for the signal
approximator case (IC, penalty, RSS, residual) \code{df.fun()} is a function that
returns the degrees of freedom of fit of %yhat = Proj_{null(D_B)} y%
}

