% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{get.states}
\alias{get.states}
\title{Function to take in output form \code{dualPathSvd2()} and returns a list of
the states after each step in algorithm, starting with NA as the first
state.}
\usage{
get.states(pathobj)
}
\description{
Function to take in output form \code{dualPathSvd2()} and returns a list of
the states after each step in algorithm, starting with NA as the first
state.
}

