% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{getadjmat.from.Dmat}
\alias{getadjmat.from.Dmat}
\title{Takes in a D matrix used in the generalized lasso and creates an adjacency
matrix (a matrix with zero entries in non-adjacent edges and 1 in adjacent
edges).}
\usage{
getadjmat.from.Dmat(Dmat)
}
\description{
Takes in a D matrix used in the generalized lasso and creates an adjacency
matrix (a matrix with zero entries in non-adjacent edges and 1 in adjacent
edges).
}

