% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{getbic.graph}
\alias{getbic.graph}
\title{Wrapper for getmodelinfo.graph() to get bic scores only Returns BIC for
 steps 0~maxsteps}
\usage{
getbic.graph(f0, y0, sigma, maxsteps, Dmat, stoptime = F, dual = T)
}
\description{
Wrapper for getmodelinfo.graph() to get bic scores only Returns BIC for
 steps 0~maxsteps
}

