% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{getbic.regression}
\alias{getbic.regression}
\title{Returns sequence of BIC of fused lasso regression model (general regressor
matrix X) The variable names 00.orig are to emphasize that while the f0 may
be from solving a ridge-penalty-added problem, what should be provided are
the original, non-augmented versions}
\usage{
getbic.regression(y0.orig, f0, sigma, maxsteps, X.orig, ginvX.orig, D.orig,
  rtol = 1e-07)
}
\description{
Returns sequence of BIC of fused lasso regression model (general regressor
matrix X) The variable names 00.orig are to emphasize that while the f0 may
be from solving a ridge-penalty-added problem, what should be provided are
the original, non-augmented versions
}

