% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{getDmat.from.adjmat}
\alias{getDmat.from.adjmat}
\title{Takes in adjacency matrix (either produced from igraph, or a matrix with
                           zero entries in non-adjacent edges and 1 in
                           adjacent edges) and produces a (sparse) D matrix
                           for usage in the generalized lasso}
\usage{
getDmat.from.adjmat(adjmat, sparseMatrix = FALSE)
}
\description{
Takes in adjacency matrix (either produced from igraph, or a matrix with
                           zero entries in non-adjacent edges and 1 in
                           adjacent edges) and produces a (sparse) D matrix
                           for usage in the generalized lasso
}

