% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{getdvec}
\alias{getdvec}
\title{get k'th dik vector, for the segment or spike test.}
\usage{
getdvec(obj, y = NULL, k, klater = k, type = c("spike", "segment",
  "hybridsegment", "general"), n)
}
\description{
get k'th dik vector, for the segment or spike test.
}

