% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{getGammat.naive}
\alias{getGammat.naive}
\title{get k'th Gammat matrix from path object \code{obj}. Note! If you're being
naive, then do whatever you want with condition.step.  # if you're using
stop times, condition.step should be the (stoptime + 2)}
\usage{
getGammat.naive(obj, y, condition.step = NULL)
}
\description{
get k'th Gammat matrix from path object \code{obj}. Note! If you're being
naive, then do whatever you want with condition.step.  # if you're using
stop times, condition.step should be the (stoptime + 2)
}

