% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{getGammat.stoprule}
\alias{getGammat.stoprule}
\title{Produces the rows to add to Gammat matrix for IC-based stopping rule in
generalized lasso signal approximation problem.}
\usage{
getGammat.stoprule(obj, y, condition.step, stoprule, sigma, consec, maxsteps,
  ebic.fac = 0.5, D, verbose = F)
}
\description{
Produces the rows to add to Gammat matrix for IC-based stopping rule in
generalized lasso signal approximation problem.
}

