% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{getGammat.stoprule.regression}
\alias{getGammat.stoprule.regression}
\title{Produces the rows to add to Gammat matrix, in the regression case requires
X.orig, ginvX.orig and D.orig which are from the pre-elastic-net regression
problem}
\usage{
getGammat.stoprule.regression(obj, y, condition.step, stoprule, sigma, consec,
  maxsteps, X.orig, ginvX.orig = NULL, D.orig, y0.orig)
}
\description{
Produces the rows to add to Gammat matrix, in the regression case requires
X.orig, ginvX.orig and D.orig which are from the pre-elastic-net regression
problem
}

