% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{getGammat.with.stoprule}
\alias{getGammat.with.stoprule}
\title{Function that takes the same arguments as the getGammat.naive(), and some
extra arguments to extract the conditioning for the stopping rule.  Returns
the final Gamma matrix and u vector with stopping rules incorporated.}
\usage{
getGammat.with.stoprule(obj, y, condition.step, stoprule, sigma, maxsteps,
  consec, ebic.fac = 0.5, type = c("tf", "graph", "regression"), X = NULL,
  D = NULL, X.orig = NULL, ginvX = NULL, D.orig = NULL,
  y0.orig = NULL, ginvX.orig = NULL, verbose = F)
}
\description{
Function that takes the same arguments as the getGammat.naive(), and some
extra arguments to extract the conditioning for the stopping rule.  Returns
the final Gamma matrix and u vector with stopping rules incorporated.
}

