% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{getmodelinfo.graph}
\alias{getmodelinfo.graph}
\title{Function to create things related to model selection, for the graph case
 Input : path object, y0, initial graph, maximum steps to take, Dmat,
 cluster size, cluster # Output: 6 p values for each possible segment test.}
\usage{
getmodelinfo.graph(f0, y0, sigma, maxsteps, Dmat, stoptime = F)
}
\description{
Function to create things related to model selection, for the graph case
 Input : path object, y0, initial graph, maximum steps to take, Dmat,
 cluster size, cluster # Output: 6 p values for each possible segment test.
}

