% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getprimal_from_dual}
\alias{getprimal_from_dual}
\title{produces primal solution from dual solution on a point in path}
\usage{
getprimal_from_dual(u, y, type = "fl1d")
}
\description{
produces primal solution from dual solution on a point in path
}

