% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getprimalmat}
\alias{getprimalmat}
\title{produces primal matrix from dual solution matrix
Example usage: getprimal(umat = fl1d(..)$u)}
\usage{
getprimalmat(umat, y, type = "fl1d")
}
\description{
produces primal matrix from dual solution matrix
Example usage: getprimal(umat = fl1d(..)$u)
}

