% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{getpvals.fl1d}
\alias{getpvals.fl1d}
\title{produce series of p-values for one sided test at each step k, wrapper for pval.fl1d()
|fl1dpath|: is from running fl1d(), and is a list that contains B, Gammat, ll, s, u
|sigma|   : is the noise of the response y = Normal(theta, sigma). We can choose to use the one we generated data from if we play god, or use an estimate
            by just taking the estimate around theta i.e. sd(y-hattheta)
NOTICE: this necessarily couples the hypothesis test with the jump made at that step!!}
\usage{
getpvals.fl1d(y, fl1d.obj, sigma)
}
\description{
produce series of p-values for one sided test at each step k, wrapper for pval.fl1d()
|fl1dpath|: is from running fl1d(), and is a list that contains B, Gammat, ll, s, u
|sigma|   : is the noise of the response y = Normal(theta, sigma). We can choose to use the one we generated data from if we play god, or use an estimate
            by just taking the estimate around theta i.e. sd(y-hattheta)
NOTICE: this necessarily couples the hypothesis test with the jump made at that step!!
}

