% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{getVs}
\alias{getVs}
\title{get k'th Vup or Vlo for user-input y, G, dik and sigma
example usage: getGammat(fl1d(..), 1y)}
\usage{
getVs(y, G, dik, sigma, u = rep(0, nrow(G)))
}
\description{
get k'th Vup or Vlo for user-input y, G, dik and sigma
example usage: getGammat(fl1d(..), 1y)
}

