% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{makesegment}
\alias{makesegment}
\title{Helper for getdvec()
|breaks|: all the breakpoints from fl1d path so far
|k|: algorithm's step to use
|klater|: later step to condition on
|n|: length of response (= length of dvec )}
\usage{
makesegment(breaks, k, klater, n)
}
\description{
Helper for getdvec()
|breaks|: all the breakpoints from fl1d path so far
|k|: algorithm's step to use
|klater|: later step to condition on
|n|: length of response (= length of dvec )
}

