% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{pval.fl1d}
\alias{pval.fl1d}
\title{produce p-value for one sided test
|y|       : is the original noisy y
|dik|     : is a vector (same length as y) with -c in i_k'th entry, 
|G|       : is the gamma vector at that step
|sigma|   : is the noise of the response y = Normal(theta, sigma).
TODO: change dik to v, without breaking everything I've ever done.}
\usage{
pval.fl1d(y, G, dik, sigma, approx = T, threshold = T,
  approxtype = c("gsell", "rob"), u = rep(0, nrow(G)))
}
\description{
produce p-value for one sided test
|y|       : is the original noisy y
|dik|     : is a vector (same length as y) with -c in i_k'th entry, 
|G|       : is the gamma vector at that step
|sigma|   : is the noise of the response y = Normal(theta, sigma).
TODO: change dik to v, without breaking everything I've ever done.
}

