% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{sbmsim}
\alias{sbmsim}
\title{Function to run post-selection tests on graphs and store p-values results
 Currently only is able to do 3-clusters
 Input : path object, y0, initial graph, maximum steps to take, Dmat, cluster size, cluster #
 Output: 6 p values for each possible segment test.}
\usage{
sbmsim(f0, y0, mygraph, maxsteps, Dmat, clustersize, nclust = 3,
  verbose = FALSE, stoptime = F)
}
\description{
Function to run post-selection tests on graphs and store p-values results
 Currently only is able to do 3-clusters
 Input : path object, y0, initial graph, maximum steps to take, Dmat, cluster size, cluster #
 Output: 6 p values for each possible segment test.
}

